KS-208 mv workflow yaml graph validation to chainlink-common  (#13235)
* prefactoring

* update common ref
