+++ .changeset/blue-pumpkins-sniff.md
@@ -1,0 +1,5 @@
+---
+"chainlink": minor
+---
+
+#internal move workflow validation to common repo
