--- core/services/workflows/testdata/fixtures/workflows/marshalling/workflow_2_spec.json
@@ -1,31 +1,0 @@
-{
-  "triggers": [
-    {
-      "id": "on_mercury_report@1",
-      "ref": "report_data",
-      "config": {}
-    }
-  ],
-  "consensus": [
-    {
-      "id": "trigger_test:aaShouldBeFirst_true:chain_ethereum:network_mainnet@2",
-      "inputs": {
-        "observations": [
-          "triggers.report_data.outputs"
-        ]
-      },
-      "config": {}
-    }
-  ],
-  "targets": [
-    {
-      "id": "write_polygon_mainnet@1",
-      "inputs": {
-        "report": [
-          "consensus.evm_median.outputs.report"
-        ]
-      },
-      "config": {}
-    }
-  ]
-}
